refactor(gui): extract voice/download tests to sibling file
Move the `#[cfg(test)] mod tests` block out of
`primer-gui/src/voice/download.rs` (958 lines) into a sibling
`src/voice/download/tests.rs`. Pure verbatim relocation — no behaviour
change, no incidental fixes.

- download.rs drops 958 -> 429 lines (-529); parent now holds only
  production command code plus a 2-line `#[cfg(test)] mod tests;`
  declaration. Child tests.rs is 525 lines.
- Non-root module: `voice/download.rs`'s submodule resolves to
  `src/voice/download/tests.rs` (new `download/` dir), same form as the
  prior `commands/voice.rs` -> `commands/voice/tests.rs` cut.
- Zero visibility churn — the tests are a descendant module reaching
  parent privates via `use super::*`.
- Feature matrix verified (speech-gated crate): default 205, speech 253
  tests pass identically pre/post move; clippy clean on both; workspace
  51 ok blocks / 0 failures.
